Add SAX 1.0 completion: AttributeList, HandlerBase, JavApiSax1Parser
Implement the remaining SAX 1.0 API surface and provide a concrete parser:

- AttributeList protocol (deprecated since SAX 2.0, replaces [String:String])
- AttributeListImpl: mutable implementation with add/remove/clear and copy constructor
- SAXNotRecognizedException: missing exception type
- HandlerBase: SAX 1.0 adapter with no-op defaults for all four handler interfaces;
  fatalError re-throws as per spec
- DocumentHandler.startElement signature changed from [String:String] to AttributeList
- JavApiSax1Parser (SwiftExtensions/XML/): concrete SAX 1.0 parser, not a Java API port
  - Apple platforms: Foundation XMLParser-backed (_SAX1FoundationDelegate bridge)
  - Linux/WASM and other non-Apple platforms: pure-Swift fallback parser
    (elements, attributes, text, CDATA, PIs, comments, entity references)
- ParserFactory: auto-registers JavApiSax1Parser as default under defaultParserName;
  no canImport(Foundation) guard needed since JavApiSax1Parser is always available
- Tests: AttributeListImpl (10), HandlerBase (13), JavApiSax1Parser (24+, Apple only)
  including large-document integration test with nested elements, attributes,
  multiple same-name children, element stacks and processing instructions

+++ Sources/JavApi/awt/toolkit/gdi/_Win32WindowHost.swift
@@ -184,7 +184,7 @@ public final class _Win32Canvas {
       // Frame/JFrame: WS_OVERLAPPEDWINDOW enthält WS_THICKFRAME — entfernen wenn nötig
       style = canResize
         ? DWORD(WS_OVERLAPPEDWINDOW)
-        : DWORD(WS_OVERLAPPEDWINDOW & ~WS_THICKFRAME)
+        : DWORD(WS_OVERLAPPEDWINDOW) & ~DWORD(WS_THICKFRAME)
     }
 
     var rect = RECT(left: LONG(x), top: LONG(y),
